Use size_t for token buffer sizes
Change NormalizeToken's out_size parameter from uint16_t& to size_t& to match sizeof usage and avoid narrowing/overflow. Update callers in ParseHexBytes, ParseEndpointValue, and ParseHexWord, and adjust the assignment in NormalizeToken (out_size = len). This removes unsafe casts and potential truncation when handling buffer sizes.

+++ sig-net-parse.cpp
@@ -489,7 +489,7 @@ static uint8_t HexNibble(char ch)
 
 static int32_t NormalizeToken(const char* text,
                               char* out_token,
-                              uint16_t& out_size,
+                              size_t& out_size,
                               bool strip_0x_prefix)
 {
     if (!text || !out_token || out_size < 2) {
@@ -519,7 +519,7 @@ static int32_t NormalizeToken(const char* text,
         return SIGNET_ERROR_BUFFER_TOO_SMALL;
     }
 
-    out_size = static_cast<uint16_t>(len);
+    out_size = len;
 
     memcpy(out_token, begin, len);
     out_token[len] = '\0';
@@ -533,7 +533,7 @@ int32_t ParseHexBytes(const char* text, uint8_t* out_bytes, uint16_t byte_count)
     }
 
     char token[256];
-    uint16_t out_size = sizeof(token);
+    size_t out_size = sizeof(token);
     int32_t norm = NormalizeToken(text, token, out_size, true);
     if (norm != SIGNET_SUCCESS) {
         return norm;
@@ -568,7 +568,7 @@ int32_t ParseTUIDHex(const char* text, uint8_t out_tuid[6])
 int32_t ParseEndpointValue(const char* text, uint16_t& endpoint_out)
 {
     char token[64];
-    uint16_t out_size = sizeof(token);
+    size_t out_size = sizeof(token);
     int32_t norm = NormalizeToken(text, token, out_size, false);
     if (norm != SIGNET_SUCCESS) {
         return norm;
@@ -596,7 +596,7 @@ int32_t ParseEndpointValue(const char* text, uint16_t& endpoint_out)
 int32_t ParseHexWord(const char* text, uint16_t& value_out)
 {
     char token[64];
-    uint16_t out_size = sizeof(token);
+    size_t out_size = sizeof(token);
     int32_t norm = NormalizeToken(text, token, out_size, false);
     if (norm != SIGNET_SUCCESS) {
         return norm;
